Mobile: tiered reconnect backoff + Terminal settings subpage
Two unrelated UX fixes that touched neighboring files.

1. Reconnect backoff. The previous ladder (500ms→4s, 6 attempts) gave
   up after ≈ 30s of failure and parked the WS in 'reconnecting'
   forever until the user killed and relaunched the app. On flaky
   networks (laptop sleep, AP-isolation cycles, brief Wi-Fi blip)
   that's a constant kill-relaunch loop.

   New ladder is tiered: snappy retries for the first ≈ 15s
   (500ms→4s) cover the common transient case, then we slow down
   (8s→60s) so we don't burn battery if the desktop is genuinely
   gone but still self-heal once the network recovers. Total runway
   to give-up is now ≈ 2 minutes across 12 attempts.
   UNREACHABLE_ATTEMPTS in connection-health.ts bumped in lockstep
   so the verdict label doesn't flip ahead of the loop pausing.

2. Terminal settings subpage. The 'When you leave the app' picker
   for auto-restore-fit lived inline on the main Settings screen,
   one row per paired host. Pulled it out to its own subpage at
   /terminal-settings with the same UI, and replaced the inline
   section with a single 'Terminal' nav row. Keeps the main
   Settings list short and gives us a home for future
   terminal-specific knobs without growing the root screen.
